delete useBlocksStore
