+++ packages/core/admin/admin/src/content-manager/components/BlocksInput/blocks/tests/Heading.test.tsx
@@ -1,5 +1,3 @@
-import * as React from 'react';
-
 import { render, screen } from '@testing-library/react';
 
 import { headingBlocks } from '../Heading';
